+++ extensions/uptime-robot/src/lib/api.ts
@@ -1,5 +1,4 @@
-import fetch from "node-fetch";
-import { ErrorResponse, SuccessResponse } from "../types";
+import { ErrorResponse, Monitor, NewMonitor, SuccessResponse } from "../types";
 import { API_BODY, API_HEADERS, API_URL } from "../config";
 
 export async function deleteMonitor(monitorId: number) {
@@ -17,4 +16,21 @@ export async function deleteMonitor(monitorId: number) {
       cause: result.error.type,
     });
   return result.monitor;
+}
+
+export async function createMonitor(newMonitor: NewMonitor) {
+  const res = await fetch(API_URL + "newMonitor", {
+    headers: API_HEADERS,
+    method: "POST",
+    body: new URLSearchParams({
+      ...API_BODY,
+      ...newMonitor
+    }).toString(),
+  });
+  const result = (await res.json()) as SuccessResponse<Monitor> | ErrorResponse;
+  if (result.stat === "fail")
+    throw new Error(result.error.message, {
+      cause: result.error.type,
+    });
+  return result;
 }