+++ extensions/uptime-robot/src/types.ts
@@ -27,11 +27,15 @@ type MonitorLog = {
     detail: string;
   };
 };
+export enum MonitorType {
+  HTTP=1,
+  PING=3
+}
 export type Monitor = {
   id: number;
   friendly_name: string;
   url: string;
-  type: number;
+  type: MonitorType;
   sub_type: string;
   keyword_type: number;
   keyword_case_type: number;
@@ -50,7 +54,7 @@ export type NewMonitor = {
   friendly_name: string;
   url: string;
   type: string;
-  interval: string;
+  interval?: string;
   timeout: string;
 };