+++ extensions/uptime-robot/CHANGELOG.md
@@ -1,5 +1,11 @@
 # UptimeRobot Changelog
 
+## [] - {PR_MERGE_DATE}
+
+- monitor type in monitors
+- create PINg
+- update deps
+
 ## [Delete Monitor + Open in Browser] - 2025-03-04
 
 - New `Action` to Delete Monitor (optimistically)